fix (logistic): solve transpose problem between the two classes
the data matrix was previously required in the shape (n_samples, n_feats) by the quadratic logistic regression, but was transposed to be fed to the mapper into the 2d space and then fed to the base class (the linear LR) which had the data in the shape (n_mapped_feats, n_samples). This forced the user to give a transpose matrix to the LR and a non-transposed one to the Qadratic LR. Rewriting the mapping function sealed the deal.